requirements.txt
streamlit==1.36.0
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.3.0
pandas==2.2.2
openpyxl==3.1.2
PyYAML==6.0.2
